backend/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
torch==2.0.1
transformers==4.36.0
sentence-transformers==2.5.1
faiss-cpu==1.7.4
google-generativeai==0.3.0
python-dotenv==1.0.0
numpy==1.24.0
requests==2.31.0 